+++ lib/web_ui/lib/src/engine/dom.dart
@@ -191,7 +191,7 @@ extension DomNodeExtension on DomNode {
   }
 
   external DomNode removeChild(DomNode child);
-  external bool? get isConnected;
+  external bool get isConnected;
   set text(String? value) =>
       js_util.setProperty<String?>(this, 'textContent', value);
   external DomNode cloneNode(bool? deep);
