+++ lib/web_ui/test/canvaskit/h5vcc_test.dart
@@ -53,7 +53,7 @@ void testMain() {
     });
 
     test('Surface acquireFrame uses getH5vccSkSurface', () {
-      final Surface surface = SurfaceFactory.instance.getSurface()!;
+      final Surface surface = SurfaceFactory.instance.getOverlay()!;
       surface.acquireFrame(ui.Size.zero);
       expect(getH5vccSkSurfaceCalledCount, 1);
